Fix issue with BST dates
Since we store Date Of Birth as a unix timestamp without a time component
it inherently has a time of 00:00AM... Which is okay, except for the fact
that during BST, we save it at 23:00AM the previous day by accident.

This forces the date saved to be 'At UTC' which I don't think is the ideal
fix but it at least gives something workable for now.