Remodel rats to allow stubbing new ones
This removes restrictions for new rats in the domain and data models -
now it is not required to set a name, sex, date of birth on a rat before
it can be stored and operated on.

This restriction may differ on other models, but the core domain will
not require it so that breeders can create rats easier. This facilitates
functionality like letting a breeder stub out a litter without
necessarily being required to sex or name the rats at that point in time
so that they can later return to the rats linked to the litter and edit
them when they have more time to do so.