Introduce litters
This is a massive change and in better circumstances, would have been
done more piecemeal and cleaner. It's not exactly in the cleanest state
right now but it's a solid amount of prototyping around what the
application's navigation might work like.

In short, this adds the ability to create, edit and delete rats and
litters, including facilities to search for existing rats (particularly
useful when linking them to a litter).

There's a number of known problems:
  - Plenty of the UI should be componentised as a lot of it is copy
    pasted everywhere.
  - The style of modelling for litters and rats is very different, and I
    think I prefer the litter side of things. Rats should ideally be
    reworked a bit to take some lessons from that.
  - There are so many missing tests - in an effort to focus on getting a
    PoC UI down and in lieu of all the backend work needed to do that,
    testing outside of manual testing took a backseat to allow for rapid
    (extremely haphazard) modifications. Now that everything has firmed
    up, we should wrap a full test harness round it before we add more
    significant functionality or rework anything, such that we don't
    cause regressions
  - KNOWN BUG: If you delete something the UI will navigate to another
    page automatically, if you then click "back" it will navigate back
    to the deleted entity's page, the initialisation will load a null
    object, it'll bind the page to that object and then crash.

wip litters

further wip

mega wip

This does everything, but is messy and without tests