Rework and test Rats page
This tidies up the rats page and separates the inline @code declaration
into a code-behind file. This also changes how some things are injected
by the runtime and sets up the test projects to be able to directly
invoke actions on these pages, so that we can test behaviour.

To that end, the Rats page is now fully tested in all but the JS Interop
which I'm not sure I care to test.