Handle where users view a deleted rat/litter
If you deleted a rat and clicked "back" then the application would fail
to load it, reference properties of it in the view, and crash. The back
button had to remain as I want to enable android users to use the
software normally.

This also removes all null build warnings by having more safety around
the nullability of rats/litters. They're declared non-nullable but
initialised to `null!` by default, and the view code now makes sure that
they have properly loaded or else it'll display some simple warning
components rather than anything which references the Rat or Litter.

I could've gone the route of declaring them nullable but it would have
been painful to guard against nullability in 90% of the logic on the
page rather than just split the view logic early and have trust in the
rest of it.