Journal db scripts
This adds journaling to the db, I should write why in an ADR but it
basically boils down to the fact I do not have access to these databases
and if I need to ever try support someone and open it up manually, I can
get a view as to what version they used.

Also adds date_of_pairing to litter since I want to make a feature with
that shortly.