Turn sqlite connection pooling off
This fixes a very annoying windows issue where we could not manipulate
the database file, manually or programmatically, while the application
was running. This is because the file was always locked as "in use by
another application" though it never seemed to have issues internally.

It turns out the Sqlite driver library implements "connection pooling"
by default, which basically means it'll open the file and keep it open
in the background until the application is shut down, regardless of
whether the IDbConnection is disposed. This makes sense, it's just not
obvious. This commit turns pooling off in all connection strings.

There's a possibility this might have unintended consequences when it
comes to performance - a future consideration may be to see if we can
manually invalidate the connection pool.

Fixes #5