Add "owned" flag to rats
This adds a very lightweight mechanism to determine if a rat in the
database is currently owned by the user. By default, all existing rats
and new rats will be considered "owned" by default, but you can update
this when editing a rat.

Also this introduces another pre-set filter on the main page to hide
unowned rats.

Very mildly related to #12