Fix bug where new litters lost date of birth
You could "lose" a date of birth if you created a new litter then
started creating new rats underneath it without manually saving the
litter. This is because on rat creation the UI will navigate away to the
rat without saving. On navigation back to the litter, it'll re-load the
litter details (which are obviously missing the date of birth now).

Fixes #2