Set litter date of birth from first searched rat
If the litter has no date of birth and you add offspring with a date of
birth already set, the system will take that as fact for the whole
litter (which makes sense if the user hasn't made an error when setting
the rat's date of birth).

Date of birth will then, as per existing logic, carry onto any rats made
from the litter.

* #31