Update CommunityToolkit
This update fixes the file saver on mac, where previously it would use a
GUID instead of the file name parameter passed in on invocation to set
the suggested file name, prompting the user to always rename the file
themselves.

Now, it correctly uses the parameter, as all other platforms do.