Ingest owners, skip user owner
Now that owners exist in the system we can identify them in the
spreadsheet, store them, and associate them to rats.

The user of the system might also identify themselves, by their name or
rattery name or even just "Me" in this field. We should skip these and
mark these rats as owned by the user. This adds a mechanism where users
can provide their own name on import to do so.