Add TEMPORARY android 13 workaround
The crux of the problem is that either MAUI, or the toolkit, or me is
doing something that doesn't quite work out with the android 13 API 33
changes to file storage access.

This change adds a line to the manifest that effectively tells the
device to fall back to an earlier SDK version. This seems to grant file
access and successfully open dialogs on android 13 devices again.

This is not a safe or permanent fix, but is just to get some existing
users some amount of control over their data. From what I read, I have
until October 2023 before this has to be on API 33.