Properly save rat before adding owner
The rat page suffers a bit of early legacy from before I had a better
model figured out (or one I prefer for now) where the UI was mostly
bound to a property bag "RatForm" model. A special page "Save" was built
to transfer values from the RatForm to the rat model before invoking
the save method, effectively serialising all changes and then storing
them.

I made a mistake when adding owners and simply invoked the rat save
method, which would save the unchanged rat. If you edited a rat, set
"owned" to false and then tried to add an owner, the "owned" status
would not be updated and the validation in the rat would fail to add
the owner to the "owned" rat.

Daft.

* #32