Add saveable settings (rattery name, pedigree config)
This moves the existing Pedigree options into settings and allows users
to persist them (so they don't have to re-type their rattery over and
over again). Also, this adds some values to allow tweaking the pedigree
in the hopes that users can resolve their own issues with the pedigree
overflowing onto a second page.

Finally, this moves the existing "Settings" (which never really was
settings) into a new "Tools" page.