Add in-memory persistence for list filter boxes
Until we add a nice distinct "All Rats" list with filter options and
persistence there, it can be quite painful for users to keep having to
remove a filter (like showing unowned rats) in some use cases (such as
setting up a system for the first time).

For now, this adds a means of persisting the last applied filter in
memory only - meaning we will lose the setting on app restart. I think
that is a reasonable experience (should persist for the duration of a
use case that necessitates it, then reverts back to the believed "most
relevant settings" on app restart).

Longer-term persistence was considered (using MAUI IPreferences) but I
do not believe it is necessary nor the best experience right now, with
the plans to introduce a separate filter for all rats.