Update QuestPDF
Some dependency of QuestPDF no longer worked with iOS simulators and this, inexplicably, caused
any build for any platform to fail when built on mac.

Updating QuestPDF has resolved this. It has also introduced a lot of errors around fonts which
I do not use. In test, this is not problematic