Disable edge-to-edge on android
The update to target SDK35 came with a problem - edge to edge rendering is now
enabled by default and the title bar renders under notification bars on most
devices. This makes it impossible to click the menu button on some phones,
meaning Basemix is completely unusable.

This opts out of edge-to-edge rendering. This should be fixed properly in due
time but other priorities are above that.