Standardise date rendering
Users are reporting wrong-format dates (american dates in the UK for example)
and it is unclear why, but during a review of date handling I noticed that
it is not consistent across the app. This introduces my own static methods
for converting to strings consistently so that there's little variation.

If problems persist, this will at least make fixes easier to deal with.